Remove AcademyApi and use HttpService
